src/requirements.txt
cloudshell-shell-core>=3.1,<3.2
cloudshell-automation-api>=7.1,<7.2
cloudshell-networking-devices>=2.3,<2.4
cloudshell-tg-trex>=1.0,<1.1